+++ packages/library/src/commonMain/kotlin/io/realm/MutableRealm.kt
@@ -119,19 +119,6 @@ class MutableRealm : BaseRealm {
         RealmInterop.realm_rollback(realmReference.dbPointer)
     }
 
-    @Deprecated("Use MutableRealm.copyToRealm() instead", ReplaceWith("io.realm.MutableRealm.copyToRealm(obj)"))
-    fun <T : RealmObject> create(type: KClass<T>): T {
-        return io.realm.internal.create(configuration.mediator, realmReference, type)
-    }
-    // Convenience inline method for the above to skip KClass argument
-    @Deprecated("Use MutableRealm.copyToRealm() instead", ReplaceWith("io.realm.MutableRealm.copyToRealm(obj)"))
-    inline fun <reified T : RealmObject> create(): T { return create(T::class) }
-
-    @Deprecated("Use MutableRealm.copyToRealm() instead", ReplaceWith("io.realm.MutableRealm.copyToRealm(obj)"))
-    fun <T : RealmObject> create(type: KClass<T>, primaryKey: Any?): T {
-        return io.realm.internal.create(configuration.mediator, realmReference, type, primaryKey)
-    }
-
     /**
      * Creates a copy of an object in the Realm.
      *