+++ test/src/androidTest/kotlin/io/realm/shared/LinkTests.kt
@@ -49,8 +49,8 @@ class LinkTests {
     fun basics() {
         val name = "Realm"
         val parent = realm.writeBlocking {
-            val parent = create(Parent::class)
-            val child = create(Child::class)
+            val parent = copyToRealm(Parent())
+            val child = copyToRealm(Child())
             child.name = name
 
             assertNull(parent.child)