+++ test/src/androidTest/kotlin/io/realm/shared/NullabilityTests.kt
@@ -46,11 +46,9 @@ class NullabilityTests {
     @Test
     fun nullability() {
         realm.writeBlocking {
-            val nullability = create(Nullability::class)
+            val nullability = copyToRealm(Nullability())
             assertNull(nullability.stringNullable)
-            // TODO We cannot verify this before implementing support for default values
-            //  https://github.com/realm/realm-kotlin/issues/106
-            // assertNonNull(nullability.stringNonNullable)
+            assertNotNull(nullability.stringNonNullable)
 
             nullability.stringNullable = "Realm"
             assertNotNull(nullability.stringNullable)